pyproject.toml
[tool.poetry]
name = "exchange_sdk"
version = "0.1.0"
description = ""
authors = ["atabakhooshangi <atabak.hooshangi.1990@gmail.com>"]
readme = "README.md"
packages = [{include = "exchange_sdk"}]

[tool.poetry.dependencies]
python = "^3.7"
python-decouple = "^3.6"
requests = "^2.28.1"


[tool.poetry.dev-dependencies]
pytest = "^6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
